refactor/#1: 그리드 뷰 파일 분리
