+++ src/container/gridViewTemplate.js
@@ -1,17 +1,17 @@
 import { GRID_ROW_SIZE, GRID_COL_SIZE, SNACK_BAR_TIME } from "../utils/constant.js";
-import { grid_view_info_entire, grid_view_info_sub, toggleArrow } from "../components/grid/gridToggle.js";
+import { grid_view_info_entire, grid_view_info_sub } from "../components/grid/gridObserver.js";
+import { toggleArrow } from "../components/grid/gridArrow.js";
 import { create } from "../utils/createElement.js";
 import { class_name } from "../utils/domClassName.js";
 import { buttonFacotry } from "../components/common/btnfactory.js";
-import { subscribe_idx_list } from "../../data/subscribeIdxList.js";
 import { createSnackBar } from "../components/common/snackBar.js";
 import { createAlert } from "../components/common/alertSubscribe.js";
-import { _sub_press_list } from "../Store.js";
 import { onClickSubBtn } from "../components/layout/mainNavEvent.js";
 
 const btnFactory = new buttonFacotry();
 
-export function createMainGrid(grid_view_info, isInit) {
+// 그리드 뷰 생성 (화살표 제외한 뷰)
+export function createMainGrid(grid_view_info, isInit, subscribe_list) {
     const $container = isInit
         ? create.div({ className: "main_news_container" })
         : document.querySelector(`${grid_view_info.getClassName()}`).querySelector(".main_news_container");
@@ -36,16 +36,18 @@ export function createMainGrid(grid_view_info, isInit) {
                 continue;
             }
 
-            const $mouse_enter_grid = create.div({
+            // 그리드 셀 호버시 나타나는 컴포넌트
+            const $mouse_hover_btn = create.div({
                 className: "mouse-enter-grid",
                 events: {
                     mouseleave: () => {
-                        $mouse_enter_grid.style.visibility = "hidden";
+                        $mouse_hover_btn.style.visibility = "hidden"; // 그리드 셀 벗어나는 경우 이벤트
                     },
                 },
             });
 
-            const is_subscribe_press = !_sub_press_list.getIdx().includes(data_list[cnt].id);
+            // 구독하기 버튼 (구독한 경우: 해지하기 버튼, 구독하지 않은 경우: 구독하기 버튼)
+            const is_subscribe_press = !subscribe_list.includes(data_list[cnt].id);
             const $subscribe_btn = btnFactory.create({
                 type: "subscribe",
                 isDefault: true,
@@ -53,15 +55,18 @@ export function createMainGrid(grid_view_info, isInit) {
                 press_id: data_list[cnt].id,
                 press_name: data_list[cnt].name,
             });
-
+            // 구독하기 버튼 클릭시 이벤트
             $subscribe_btn.setEvents({
                 click: () => {
                     const $snack_bar = createSnackBar();
 
                     const $id = $subscribe_btn.getId();
                     if (!$subscribe_btn.getIsSubscribe()) {
+                        // 해지하기 버튼 클릭시 구독해지 알림
                         $container.appendChild(createAlert($subscribe_btn.getName(), $id));
                     } else {
+                        // 구독하기 버튼 클릭시 구독한 언론사 리스트에 추가 및 스낵바
+                        // 스낵바 5초간 유지 후 내가 구독한 언론사로 이동
                         $subscribe_btn.changeMode();
                         _sub_press_list.addState($id);
                         document.querySelector(".main_news_container").appendChild(createSnackBar());
@@ -72,13 +77,13 @@ export function createMainGrid(grid_view_info, isInit) {
                     }
                 },
             });
-            $mouse_enter_grid.appendChild($subscribe_btn.getButton());
+            $mouse_hover_btn.appendChild($subscribe_btn.getButton());
 
             const $list_item = create.li({
                 className: list_class_name,
                 events: {
                     mouseenter: () => {
-                        $mouse_enter_grid.style.visibility = "visible";
+                        $mouse_hover_btn.style.visibility = "visible";
                     },
                 },
             });
@@ -87,7 +92,7 @@ export function createMainGrid(grid_view_info, isInit) {
                 className: "news_data_img",
                 attributes: { src: data_list[cnt++].press_light_src },
             });
-            $list_item.append($list_img, $mouse_enter_grid);
+            $list_item.append($list_img, $mouse_hover_btn);
             $list.appendChild($list_item);
         }
         $container.appendChild($list);
@@ -112,17 +117,17 @@ function createGridArrowBtn(btnFactory, isRight, is_subscribe, grid_view_info) {
     });
 }
 
-function createGridView(is_subscribe, grid_view_info, is_init) {
+function createGridView(is_subscribe, grid_view_info) {
     const $container = document.querySelector(`.grid-${is_subscribe}`);
     $container.append(
         createGridArrowBtn(btnFactory, false, is_subscribe, grid_view_info).getButton(),
-        createMainGrid(grid_view_info, is_init),
+        createMainGrid(grid_view_info, true, []),
         createGridArrowBtn(btnFactory, true, is_subscribe, grid_view_info).getButton()
     );
     toggleArrow(grid_view_info);
 }
 
 export function gridView(is_init) {
-    createGridView(class_name.ENTIRE, grid_view_info_entire, is_init);
-    createGridView(class_name.SUBSCRIBE, grid_view_info_sub, is_init);
+    createGridView(class_name.ENTIRE, grid_view_info_entire);
+    createGridView(class_name.SUBSCRIBE, grid_view_info_sub);
 }
